% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/budget.R
\docType{data}
\name{budget}
\alias{budget}
\title{budget}
\format{
A tibble with 62 rows and 4 variables:
\describe{
  \item{Category}{dbl}
  \item{Budget}{dbl} 
  \item{Period2}{dbl}
  \item{Period4}{dbl}
  
}
}
\usage{
budget
}
\description{
budget
}
\keyword{datasets}
