% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendar.R
\docType{data}
\name{calendar}
\alias{calendar}
\title{calendar}
\format{
A tibble with 2010 rows and 17 variables:
\describe{
  \item{ID}{dbl}
  \item{Date}{chr} 
  \item{DayNumberOfWeek}{dbl}
  \item{DayName}{chr}
  \item{MonthName}{chr}
  \item{MonthNumberOfYear}{dbl}
  \item{DayNumberOfYear}{dbl}
  \item{WeekNumberOfYear}{dbl}
  \item{CalendarQuarter}{dbl}
  \item{CalendarYear}{dbl}
  \item{FiscalYear}{dbl}
  \item{FiscalSemester}{dbl}
  \item{FiscalQuarter}{dbl}
  \item{FinMonthNumberOfYear}{dbl}
  \item{DayNumberOfMonth}{dbl}
  \item{Period}{dbl}
  
}
}
\usage{
calendar
}
\description{
calendar
}
\keyword{datasets}
