\name{CustomerCategoricals}



\title{CustomerCategoricals}
\description{
  CustomerCategorical schema
}
\usage{
CustomerCategoricals()


\details{
  The purpose of this functions is to provide dataset schema to AdventureWorks 
  customers.

  \code{del.test()} will remove the menu.
}

\examples{
TrainingPartition = CustomerCategoricals(TrainingPartition) 
}
\keyword{misc}
