% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/territory.R
\docType{data}
\name{territory}
\alias{territory}
\title{territory}
\format{
A tibble with 10 rows and 4 variables:
\describe{
  \item{Territory Key}{dbl unique ID for each territory}
  \item{Region}{chr region} 
  \item{Country}{chr country}
  \item{Group}{dbl continent}
  
}
}
\usage{
territory
}
\description{
territory
}
\keyword{datasets}
